Return @Nullable Integer for NTP customized primary colors
Avoid sentinel collision between Color.WHITE (-1) and
NtpThemeColorInfo.COLOR_NOT_SET (-1) by returning @Nullable @ColorInt
Integer from SharedPreferences primary color getters:
- getCustomizedPrimaryColorFromSharedPreference()
- getCustomizedPrimaryColorDarkFromSharedPreference()
- getDailyRefreshCustomizedPrimaryColorFromSharedPreference()
- NtpThemeDailyRefreshManager.getNtpThemeColorForThemeCollection()

Also remove obsolete NtpThemeColorInfo.COLOR_NOT_SET constant and reset
COLOR_FROM_HEX initialization to DEFAULT when primary color is missing.

TAG=agy
CONV=d64f8cf1-1f94-4f37-b627-a6359e608c29

Bug: 384012062
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8423051
Commit-Queue: Jenna Himawan <jhimawan@google.com>
Cr-Commit-Position: refs/heads/main@{#1701491}

+++ chrome/browser/ntp_customization/junit/src/org/chromium/chrome/browser/ntp_customization/theme/theme_collections/NtpThemeCollectionManagerUnitTest.java
@@ -133,7 +133,7 @@ public class NtpThemeCollectionManagerUnitTest {
         @ColorInt Integer primaryColor = captor.getValue().getPrimaryColor();
         assertNotNull(primaryColor);
         assertEquals(
-                primaryColor.intValue(),
+                primaryColor,
                 NtpCustomizationUtils.getCustomizedPrimaryColorFromSharedPreference());
     }
